Refactor make_boxed_from_unboxed_functor.h
Summary:

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
